Maps/Warnings: Reduce the amount of warnings/err...

+++ src/tools/ToolsClassic/mmaps_generator/TerrainBuilder.cpp
@@ -191,7 +191,7 @@ namespace MMAP
                 count += fread(v9, sizeof(uint8), V9_SIZE_SQ, mapFile);
                 count += fread(v8, sizeof(uint8), V8_SIZE_SQ, mapFile);
                 if (count != expected)
-                    printf("TerrainBuilder::loadMap: Failed to read some data expected %d, read %d\n", expected, count);
+                    printf("TerrainBuilder::loadMap: Failed to read some data expected %d, read %llu\n", expected, count);
 
                 heightMultiplier = (hheader.gridMaxHeight - hheader.gridHeight) / 255;
 
@@ -209,7 +209,7 @@ namespace MMAP
                 count += fread(v9, sizeof(uint16), V9_SIZE_SQ, mapFile);
                 count += fread(v8, sizeof(uint16), V8_SIZE_SQ, mapFile);
                 if (count != expected)
-                    printf("TerrainBuilder::loadMap: Failed to read some data expected %d, read %d\n", expected, count);
+                    printf("TerrainBuilder::loadMap: Failed to read some data expected %d, read %llu\n", expected, count);
 
                 heightMultiplier = (hheader.gridMaxHeight - hheader.gridHeight) / 65535;
 
@@ -221,11 +221,11 @@ namespace MMAP
             }
             else
             {
-                int count = 0;
+                size_t count = 0;
                 count += fread(V9, sizeof(float), V9_SIZE_SQ, mapFile);
                 count += fread(V8, sizeof(float), V8_SIZE_SQ, mapFile);
                 if (count != expected)
-                    printf("TerrainBuilder::loadMap: Failed to read some data expected %d, read %d\n", expected, count);
+                    printf("TerrainBuilder::loadMap: Failed to read some data expected %d, read %llu\n", expected, count);
             }
 
             // hole data
